Modified NetKAT JSON parser for Felix.
Nate is hosting NetKAT encoded Topology Zoo policies and topoologies
online [1]. The policies are encoded as JSON files, but the JSON
encoding is a bit out of sync with the current (at least as of this
commit) JSON parser in the frenetic repo [2]. Specifically, the topology
zoo policies encode switch values as integers like this:

        {
            "header": "switch",
            "type": "test",
            "value": 2
        }

but the JSON parser expects the values to be strings. I've modified the
parser is a pretty wonky ad-hoc way to get it to work with the Topology
Zoo files.

I've done something similar to this before [3]. This is just an updated
version of that.

[1]: http://www.cs.cornell.edu/~jnfoster/tmp/topozoo.tar.gz
[2]: https://github.com/frenetic-lang/frenetic/blob/a3658c86eb3d5f60c96ef35e8990ebb48dcb8729/lib/Frenetic_NetKAT_Json.ml
[3]: https://github.com/mwhittaker/frenetic/tree/json